requirements.txt
requests
mysql-connector-python
fastapi
uvicorn
python-dotenv
httpx
# scheduler 排程器
apscheduler
